Allow configuring when links should be enabled
Links can now be configured with whether they should be enabled before,
during or after the item.

Sometimes streaming links aren't a link to the specific item, but to a
stream for the room the item is in. People can get confused if the
click on a "Click to view stream" link for an item that isn't currently
on and see a different item. To avoid this confusion, we can now
configure the streaming links to only be enabled during the item.

Similarly, replays aren't typically available until after the item has
finished.

If no `WHEN` is configured, it is assumed to be always enabled.