Allow object root for program.json and people.json
In order to be able to add metadata about the files, such as version and
publish date, in a JSON compatible format the root needs to be an object
instead of an array.

This patch keeps compatibility with the old array root by testing if the
root is an array or not.