Use Uppecase only compressed ids value for QR code
Rationale:
Some cons (such as Glasgow2024) use UUIDs which are repeatative and
large and somewhat redundant. As a result their QR codes can be oversized,
so it is worth trying to match the entropy to QR codes better.

By compressing the ids and base32 formatting them we can put all
the entropy in 5bit tokens which are representin in 5.5 bits in the
QR code.

Todo:
   - Refactor ItembyIdList and GzItembyIdList to share common code.
   - Change qr library to one which uses the 5.5 bit ALPHANUM encoding
     scheme