Hide a11y labels without making them display: none
Screen readers tend to honor display: none, which means they end up not
reading them out. Instead, use other techniques to visually hide the
labels while still keeping in the in the DOM and "displayed". This
allows screen readers to more reliably read them.

It is using labels rather than aria-label because support for aria-label
is still not where it should be.