Don't render collapsed details panels
Currently because they are in the DOM and visible, just with 0 height,
both keyboard navigation and screen readers will access the elements
inside of them. This means you are tabbing the focus between elements
you can't see, and for screen readers it is reading out all of the
details for every item.

By not rendering the collapsed panels, both can now skip through the
headings of collapsed items just like a sighted mouse user can.